Revert "Go back to float precision vector storage."
This reverts commit f4c24ed21290eb914d09abce3e510f1e54fbcdf5.

+++ h2o-algos/src/main/java/hex/DataInfo.java
@@ -401,14 +401,14 @@ public class DataInfo extends Keyed {
    * @param in input values
    * @param out output values (can be the same as input)
    */
-  public final void unScaleNumericals(float[] in, float[] out) {
+  public final void unScaleNumericals(double[] in, double[] out) {
     if (_nums == 0) return;
     assert (in.length == out.length);
     assert (in.length == fullN());
     for (int k=numStart(); k < fullN(); ++k) {
-      double m = _normMul == null ? 1 : _normMul[k-numStart()];
-      double s = _normSub == null ? 0 : _normSub[k-numStart()];
-      out[k] = (float) (in[k] / m + s);
+      double m = _normMul == null ? 1f : _normMul[k-numStart()];
+      double s = _normSub == null ? 0f : _normSub[k-numStart()];
+      out[k] = in[k] / m + s;
     }
   }
